requirements.txt
torch~=2.0.0
torchvision~=0.15.1
timm~=0.6.13
numpy~=1.24.2
Pillow~=9.5.0
loguru~=0.7.0
einops~=0.6.0
piq~=0.7.0
tqdm~=4.65.0
pyiqa~=0.1.6.3
opencv-python~=4.7.0.72
matplotlib~=3.7.1
scipy~=1.10.1
lpips~=0.1.4